mv string function to string_utility and rm old file
